requirement.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
configparser==5.3.0
Django==4.1.7
djangorestframework==3.14.0
greenlet==2.0.2
idna==3.4
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
setuptools==59.6.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.4
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.14
wheel==0.37.1